+++ src/main/kotlin/dev/furq/holodisplays/HoloDisplays.kt
@@ -25,7 +25,7 @@ import java.util.concurrent.Executors
 class HoloDisplays : ModInitializer {
     companion object {
         const val MOD_ID = "HoloDisplays"
-        const val VERSION = "0.4.3"
+        const val VERSION = "0.4.5"
         val LOGGER: Logger = LoggerFactory.getLogger(MOD_ID)
 
         var SERVER: MinecraftServer? = null